+++ packages/ast-utilities/src/markdown/transforms/index.ts
@@ -1,1 +1,2 @@
 export {default as addReleaseToChangelog} from './addReleaseToChangelog';
+export {default as addBaseLinkUrl} from './addBaseLinkUrl';