+++ packages/ast-utilities/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Added
+
+- New transform `addBaseLinkUrl()` adds a base URL string to web links [[#2105])](https://github.com/Shopify/quilt/pull/2105)]
 
 ## 1.2.0 - 2021-11-24